extract to multi files
