+++ cmd/status-cli/.gitignore
@@ -1,2 +1,3 @@
 log
 *.log
+status-cli
